Progress on education/work history
Still working on a way to fix the latest issue. I've identified that the issue is that it fails to successfully fill out the field of study, so then it continues and attempts to click on the resume part of the application. I now have the ability to check if a "No matches" is displayed on the screen. I implemented an input.clear(), which didn't clear the input form. Additional updates are to come.